refactor: arrange project to be able to share assets, in preperation for adding benchmarks
in order to guage performance improvements it is important to have benchmarks which measure. move all non source code to test directory. add benchmark project with some initial benchmarks. add assets as links to console and benchmark project to ensure it works via debug and release, because benchmarks are executed in a different folder by the runner. leave actual project in current directory to ensure git history is not broken.